Ajustes al vector auxiliar del kernel bench Sobel
Se usa un tamaño ajustado de los vectores con la variable DCOLS y
balancear el desempeño de la GPU y la CPU respecto a FPGA

+++ benchsuite/mersenne.hpp
@@ -53,7 +53,7 @@ public:
   int _total_size;
 #pragma GCC diagnostic ignored "-Wignored-attributes"
   vector<int,vecAllocator<int>> _input_seed;  // image
-  vector<int,vecAllocator<int>> _out;  // blurred
+  vector<float,vecAllocator<float>> _out;  // blurred
 #pragma GCC diagnostic pop
 };
